Add -Werror, and more flags (#73)
* Add -Werror, and more flags

* Remove deprecated clang/gcc compilers from GitHub Actions

* Update upload-artifact to @v4, remove more gcc/clang8 lines

* Fix Clang warnings, update libc/libstdc package

* Change libstdc++-dev to libstdc++-13-dev

* Update mingw main.yml

* Fix git/cmake/toolchain not being mingw-w64 packages

* Fix cmake and toolchain packages

* Fix minor mingw invocation details

* Revert build.sh changes

* Add clang to mingw-w64

* Correct link flags

* Fix build.sh line-endings

* Fix -fsanitize=leak not supported on msvc/mingw

* Disable clang-scan-deps

* Fix rest of fsanitize flags

* Update to use $<BOOL:>>

* Fix win32 specific warning

* Fix Windows specific warning, again

* Fix -fsanitize=address on mingw

* Revert CMake minimum version, remove libstdc++-<version>-dev

+++ src/gfx_old.cc
@@ -829,7 +829,7 @@ int gfx_drawtext_partial(CoreState *cs, std::string text, int pos, std::size_t l
 
     for(i = pos; i < text.size() && i < len; i++)
     {
-        int thisLineLength = lines[linefeeds].size() - (linefeeds != lines.size() - 1 ? 1 : 0);
+        size_t thisLineLength = lines[linefeeds].size() - (linefeeds != lines.size() - 1 ? 1 : 0);
 
         if(thisLineLength > fmt->wrap_length)
         {
